+++ client/src/assets/hero.png
