+++ package.json
@@ -1,0 +1,32 @@
+{
+    "name": "distrubution-finder",
+    "version": "1.0.0",
+    "description": "",
+    "main": "server/server.js",
+    "scripts": {
+      "start": "node server/server.js",
+      "develop": "concurrently \"cd server && npm run watch\" \"cd client && npm start\"",
+      "install": "cd server && npm i && cd ../client && npm i",
+      "build": "cd client && npm run build"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "devDependencies": {
+      "concurrently": "^5.1.0"
+    },
+    "dependencies": {
+      "apollo-client": "^2.6.10",
+      "apollo-server-express": "^3.12.0",
+      "bcrypt": "^5.1.0",
+      "bcryptjs": "^2.4.3",
+      "dotenv": "^16.0.3",
+      "graphql": "^15.8.0",
+      "if-env": "^1.0.4",
+      "jsonwebtoken": "^9.0.0",
+      "mongo": "^0.1.0",
+      "mongoose": "^7.1.0",
+      "mutation": "^0.0.1",
+      "nodemon": "^2.0.22"
+    }
+}
