+++ server/package.json
@@ -1,0 +1,27 @@
+{
+    "name": "server",
+    "version": "1.0.0",
+    "description": "",
+    "main": "server.js",
+    "scripts": {
+      "start": "node server.js",
+      "watch": "nodemon server.js"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+      "apollo-server-express": "^3.12.0",
+      "bcrypt": "^5.1.0",
+      "bcryptjs": "^2.4.3",
+      "dotenv": "^16.0.3",
+      "express": "^4.17.1",
+      "graphql": "^16.6.0",
+      "jsonwebtoken": "^9.0.0",
+      "mongo": "^0.1.0",
+      "mongoose": "^7.1.0"
+    },
+    "devDependencies": {
+      "nodemon": "^2.0.3"
+    }
+  }
